[DTS2018070600032] Added null check on filestatus <Remove this temp fix after merging https://github.com/apache/carbondata/pull/2465 > Modification_Reason: While doing any operation on Carbon File if file status acquiring throws exception then it is logged and not thrown again therefore if any further operation on that object required filestatus then it can throw NPE. Modification_Content: Add null check and throw proper exception. Modification_Person: Kunal Kapoor Review_Person: Kanaka kumar avvaru
Refer http://code.huawei.com/BigdataAI_Kernel/Carbon/merge_requests/215

+++ core/src/main/java/org/apache/carbondata/core/datastore/filesystem/S3CarbonFile.java
@@ -61,6 +61,9 @@ public class S3CarbonFile extends HDFSCarbonFile {
   @Override
   public boolean renameForce(String changeToName) {
     FileSystem fs;
+    if (fileStatus == null) {
+      throw new RuntimeException("Unable to get file status, Please check logs for details.");
+    }
     try {
       deleteFile(changeToName, FileFactory.getFileType(changeToName));
       fs = fileStatus.getPath().getFileSystem(hadoopConf);
@@ -123,6 +126,9 @@ public class S3CarbonFile extends HDFSCarbonFile {
 
   @Override
   public CarbonFile getParentFile() {
+    if (fileStatus == null) {
+      throw new RuntimeException("Unable to get file status, Please check logs for details.");
+    }
     Path parent = fileStatus.getPath().getParent();
     return null == parent ? null : new S3CarbonFile(parent, hadoopConf);
   }
